# Separate get scalar function