Separate get scalar logic for `getExtendedPublicKey`