@paulmillr 2025-03-31
How much of a performance gain are we talking about? A comparison before/after would be useful. If it's negligible then we're optimizing the wrong part.