@kigawas 2025-03-31
@paulmillr Around 2% performance improvement, and it reduces memory footprints as well (since we can save a call to `toAffineMemo`.

I ran with `(npm run build:clean || true) && npm run build && npm run bench:install && node benchmark/curves.js`:

## Before (main branch)
```
$ (npm run build:clean || true) && npm run build && npm run bench:install && node benchmark/curves.js

> @noble/curves@1.8.1 build:clean
> rm {.,esm,abstract,esm/abstract}/*.{js,d.ts,d.ts.map,js.map} 2> /dev/null


> @noble/curves@1.8.1 build
> tsc && tsc -p tsconfig.cjs.json


> @noble/curves@1.8.1 bench:install
> cd benchmark; npm install; npm install .. --install-links

...

ed25519
init 18ms
getPublicKey x 10,207 ops/sec @ 97μs/op ± 1.13% (82μs..1ms)
sign x 5,104 ops/sec @ 195μs/op
verify x 1,142 ops/sec @ 875μs/op

ed448
init 40ms
getPublicKey x 4,494 ops/sec @ 222μs/op
sign x 2,128 ops/sec @ 469μs/op
verify x 465 ops/sec @ 2ms/op
...

```

## After
```
$ (npm run build:clean || true) && npm run build && npm run bench:install && node benchmark/curves.js

> @noble/curves@1.8.1 build:clean
> rm {.,esm,abstract,esm/abstract}/*.{js,d.ts,d.ts.map,js.map} 2> /dev/null


> @noble/curves@1.8.1 build
> tsc && tsc -p tsconfig.cjs.json


> @noble/curves@1.8.1 bench:install
> cd benchmark; npm install; npm install .. --install-links

...

ed25519
init 18ms
getPublicKey x 10,599 ops/sec @ 94μs/op ± 1.66% (78μs..3ms)
sign x 5,189 ops/sec @ 192μs/op
verify x 1,096 ops/sec @ 911μs/op ± 3.45% (773μs..6ms)

ed448
init 39ms
getPublicKey x 4,616 ops/sec @ 216μs/op
sign x 2,167 ops/sec @ 461μs/op
verify x 459 ops/sec @ 2ms/op

...
```

## After (another try)

```
ed25519
init 18ms
getPublicKey x 10,360 ops/sec @ 96μs/op ± 1.07% (81μs..1ms)
sign x 5,203 ops/sec @ 192μs/op
verify x 1,147 ops/sec @ 871μs/op

ed448
init 39ms
getPublicKey x 4,557 ops/sec @ 219μs/op
sign x 2,162 ops/sec @ 462μs/op
verify x 442 ops/sec @ 2ms/op ± 1.85% (2ms..9ms)
```